+++ packages/cli/src/index.ts
@@ -1,13 +1,19 @@
 import chalk from 'chalk'
 import { Command } from 'commander'
 import { dirname } from 'path'
-import { generator, runGenerator, getContext, FeathersBaseContext, version } from '@feathersjs/generators'
+import {
+  generator,
+  runGenerator,
+  getContext,
+  FeathersBaseContext,
+  version
+} from '@shiftsmartinc/feathers-generators'
 
 export * from 'commander'
 export { chalk }
 
 export const commandRunner = (name: string) => async (options: any) => {
-  const folder = dirname(require.resolve('@feathersjs/generators'))
+  const folder = dirname(require.resolve('@shiftsmartinc/feathers-generators'))
   const ctx = getContext<FeathersBaseContext>({
     ...options
   })
@@ -25,7 +31,7 @@ export const program = new Command()
 
 program
   .name('feathers')
-  .description('The Feathers command line interface 🕊️')
+  .description('The Feathers command line interface 🕊️, modified for Shiftsmart')
   .version(version)
   .showHelpAfterError()
 
@@ -43,6 +49,13 @@ generate
   .option('--name <name>', 'The service name')
   .option('--path <path>', 'The path to register the service on')
   .option('--type <type>', 'The service type (knex, mongodb, custom)')
+  .option('--auth <authentication>', 'does the service require authentication? (y, n)', (value) =>
+  !value ? undefined : /^(y|true|1)/i.test(value)
+  )
+  .option('--schema <schema>', 'The schema type (typebox, json, false)', (value: string) =>
+  /false|none|0/i.test(value) ? false : value
+  )
+  .option('--singular', 'Do not pluralize the service name')
   .option('--templates <templatesRoot>', 'Path to a custom templates folder (optional)')
   .action(commandRunner('service'))