+++ packages/generators/src/app/templates/app.test.tpl.ts
@@ -2,9 +2,11 @@ import { generator, toFile } from '@feathershq/pinion'
 import { renderSource } from '../../commons'
 import { AppGeneratorContext } from '../index'
 
-const template = ({
-  lib
-}: AppGeneratorContext) => /* ts */ `// For more information about this file see https://dove.feathersjs.com/guides/cli/app.test.html
+const template = ({ lib }: AppGeneratorContext) => /* ts */ `/**
+ * @external https://feathersjs.com/guides/cli/app.test.html
+ * @description For more information about this file see the link above.
+ */
+
 import assert from 'assert'
 import axios from 'axios'
 import type { Server } from 'http'