+++ packages/generators/src/app/templates/client.tpl.ts
@@ -2,14 +2,22 @@ import { generator, toFile, when } from '@feathershq/pinion'
 import { renderSource } from '../../commons'
 import { AppGeneratorContext } from '../index'
 
-const template = ({
-  name,
-  language
-}: AppGeneratorContext) => /* ts */ `// For more information about this file see https://dove.feathersjs.com/guides/cli/client.html
-import { feathers } from '@feathersjs/feathers'
+const template = ({ name, language }: AppGeneratorContext) => /* ts */ `/**
+ * @external https://feathersjs.com/guides/cli/client.html
+ * @description For more information about this file see the link above.
+ */
+
 import type { TransportConnection, Application } from '@feathersjs/feathers'
-import authenticationClient from '@feathersjs/authentication-client'
 import type { AuthenticationClientOptions } from '@feathersjs/authentication-client'
+
+import { feathers } from '@feathersjs/feathers'
+import authenticationClient from '@feathersjs/authentication-client'
+
+// #region Service Type Imports: These services will be made available within the client.
+// #endregion Service Type Imports
+
+// #region Service Type Exports: These are used to type the generated client.
+// #endregion Service Type Exports
 
 export interface Configuration {
   connection: TransportConnection<ServiceTypes>
@@ -24,7 +32,7 @@ export type ClientApplication = Application<ServiceTypes, Configuration>
  * 
  * @param connection The REST or Socket.io Feathers client connection
  * @param authenticationOptions Additional settings for the authentication client
- * @see https://dove.feathersjs.com/api/client.html
+ * @see https://feathersjs.com/api/client.html
  * @returns The Feathers client application
  */
 export const createClient = <Configuration = any> (