+++ packages/generators/src/service/templates/schema.json.tpl.ts
@@ -2,26 +2,23 @@ import { generator, toFile, when } from '@feathershq/pinion'
 import { fileExists, localTemplate, renderSource } from '../../commons'
 import { ServiceGeneratorContext } from '../index'
 
-const authFieldsTemplate = (authStrategies: string[]) =>
-  authStrategies
-    .map((name) =>
-      name === 'local'
-        ? `    email: { type: 'string' },
-    password: { type: 'string' }`
-        : `    ${name}Id: { type: 'string' }`
-    )
-    .join(',\n')
-
 const template = ({
   camelName,
   upperName,
+  singluarCamelName,
+  singluarUpperName,
+  fileName,
   relative,
   authStrategies,
   isEntityService,
   type,
   cwd,
   lib
-}: ServiceGeneratorContext) => /* ts */ `// For more information about this file see https://dove.feathersjs.com/guides/cli/service.schemas.html
+}: ServiceGeneratorContext) => /* ts */ `/**
+ * @external https://feathersjs.com/guides/cli/service.schemas.html
+ * @description For more information about this file see the link above.
+ */
+
 import { resolve, getValidator, querySyntax } from '@feathersjs/schema'${
   type === 'mongodb'
     ? `
@@ -37,26 +34,15 @@ import { dataValidator, queryValidator } from '${relative}/${
 }validators'
 
 // Main data model schema
-export const ${camelName}Schema = {
-  $id: '${upperName}',
-  type: 'object',
-  additionalProperties: false,
-  required: [ '${type === 'mongodb' ? '_id' : 'id'}', ${localTemplate(authStrategies, `'email'`, `'text'`)} ],
-  properties: {
-    ${type === 'mongodb' ? `_id: ObjectIdSchema(),` : `id: { type: 'number' },`}
-    ${
-      isEntityService
-        ? authFieldsTemplate(authStrategies)
-        : `
-    text: { type: 'string' }`
-    }
-  }
-} as const
-export type ${upperName} = FromSchema<typeof ${camelName}Schema>
-export const ${camelName}Validator = getValidator(${camelName}Schema, dataValidator)
-export const ${camelName}Resolver = resolve<${upperName}, HookContext>({})
 
-export const ${camelName}ExternalResolver = resolve<${upperName}, HookContext>({
+import { ${singluarCamelName} as ${camelName}Schema } from './${fileName}.schema.gen'
+export { ${camelName}Schema };
+export type ${singluarUpperName} = FromSchema<typeof ${camelName}Schema>
+
+export const ${camelName}Validator = getValidator(${camelName}Schema, dataValidator)
+export const ${camelName}Resolver = resolve<${singluarUpperName}, HookContext>({})
+
+export const ${camelName}ExternalResolver = resolve<${singluarUpperName}, HookContext>({
   ${localTemplate(
     authStrategies,
     `// The password should never be visible externally