+++ packages/generators/src/service/templates/test.tpl.ts
@@ -2,19 +2,19 @@ import { generator, toFile } from '@feathershq/pinion'
 import { renderSource } from '../../commons'
 import { ServiceGeneratorContext } from '../index'
 
-const template = ({
-  relative,
-  lib,
-  path
-}: ServiceGeneratorContext) => /* ts */ `// For more information about this file see https://dove.feathersjs.com/guides/cli/service.test.html
-import assert from 'assert'
+const template = ({ relative, lib, path }: ServiceGeneratorContext) => /* ts */ `/**
+ * @external https://feathersjs.com/guides/cli/service.test.html
+ * @description For more information about this file see the link above.
+ */
+
+import { describe, expect, test } from 'vitest';
 import { app } from '../${relative}/${lib}/app'
 
 describe('${path} service', () => {
-  it('registered the service', () => {
-    const service = app.service('${path}')
+  test('registered the service', () => {
+    const service = app.service('${path}');
 
-    assert.ok(service, 'Registered the service')
+    expect(service, 'Registered the service').toBeTruthy();
   })
 })
 `