+++ packages/generators/src/service/type/mongodb.tpl.ts
@@ -5,11 +5,16 @@ import { ServiceGeneratorContext } from '../index'
 export const template = ({
   className,
   upperName,
+  singluarUpperName,
   schema,
   fileName,
   kebabPath,
   relative
-}: ServiceGeneratorContext) => /* ts */ `// For more information about this file see https://dove.feathersjs.com/guides/cli/service.class.html#database-services
+}: ServiceGeneratorContext) => /* ts */ `/**
+ * @external https://feathersjs.com/guides/cli/service.class.html#database-services
+ * @description For more information about this file see the link above.
+ */
+
 import type { Params } from '@feathersjs/feathers'
 import { MongoDBService } from \'@feathersjs/mongodb\'
 import type { MongoDBAdapterParams, MongoDBAdapterOptions } from \'@feathersjs/mongodb\'
@@ -18,7 +23,7 @@ import type { Application } from '${relative}/declarations'
 ${
   schema
     ? `import type {
-  ${upperName},
+  ${singluarUpperName},
   ${upperName}Data,
   ${upperName}Patch,
   ${upperName}Query
@@ -32,18 +37,22 @@ type ${upperName}Query = any
 `
 }
 
-export type { ${upperName}, ${upperName}Data, ${upperName}Patch, ${upperName}Query }
+export type { ${singluarUpperName}, ${upperName}Data, ${upperName}Patch, ${upperName}Query }
 
 export interface ${upperName}Params extends MongoDBAdapterParams<${upperName}Query> {
 }
 
-// By default calls the standard MongoDB adapter service methods but can be customized with your own functionality.
+/**
+ * By default calls the standard MongoDB adapter service methods but can be
+ * customized with your own functionality.
+ */
 export class ${className}<ServiceParams extends Params = ${upperName}Params>
-  extends MongoDBService<${upperName}, ${upperName}Data, ${upperName}Params, ${upperName}Patch> {
+  extends MongoDBService<${singluarUpperName}, ${upperName}Data, ServiceParams, ${upperName}Patch> {
 }
 
 export const getOptions = (app: Application): MongoDBAdapterOptions => {
   return {
+    id: 'uuid',
     paginate: app.get('paginate'),
     Model: app.get('mongodbClient').then(db => db.collection('${kebabPath}'))
   }